requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
filelock==3.12.4
fsspec==2023.9.2
huggingface-hub==0.17.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
packaging==23.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.8.0
urllib3==2.0.5
